backend/requirements.txt
asgiref==3.2.7            # via django
attrs==19.3.0             # via pytest
certifi==2020.4.5.1       # via requests
chardet==3.0.4            # via requests
django==3.0.5             # via -r requirements.in, djangorestframework
djangorestframework==3.11.0  # via -r requirements.in
idna==2.9                 # via requests
importlib-metadata==1.6.0  # via pluggy, pytest
more-itertools==8.2.0     # via pytest
packaging==20.3           # via pytest
pillow==8.3.1
pluggy==0.13.1            # via pytest
py==1.8.1                 # via pytest
pyparsing==2.4.7          # via packaging
pytest-django==3.9.0
pytest==5.4.1             # via pytest-django
pytz==2019.3              # via django
requests==2.23.0
six==1.14.0               # via packaging
sorl-thumbnail==12.6.3
sqlparse==0.3.1           # via django
urllib3==1.25.9           # via requests
wcwidth==0.1.9            # via pytest
zipp==3.1.0               # via importlib-metadata
mixer==7.1.2
djangorestframework-simplejwt==4.7.1
django-cors-headers==3.8.0
django-filter==2.4.0
djoser==2.1.0
psycopg2-binary==2.8.5
drf-extra-fields==3.1.1
django-enumfield==2.0.2
python-dotenv==0.19.0
isort==5.9.3
gunicorn==20.0.4